types: Moved SoundTransform from core to types
